+++ src/main/java/kitchenpos/menu/domain/MenuValidator.java
@@ -2,6 +2,7 @@ package kitchenpos.menu.domain;
 
 import kitchenpos.common.Price;
 import kitchenpos.exception.CannotFindException;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Component;
