+++ src/test/java/kitchenpos/ordertable/application/OrderServiceTest.java
@@ -4,6 +4,7 @@ import kitchenpos.common.Price;
 import kitchenpos.common.Quantity;
 import kitchenpos.exception.CannotFindException;
 import kitchenpos.menu.domain.*;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.*;
 import kitchenpos.order.dto.OrderRequest;
