+++ src/test/java/kitchenpos/ordertable/ui/OrderControllerTest.java
@@ -3,8 +3,9 @@ package kitchenpos.ordertable.ui;
 import kitchenpos.common.Price;
 import kitchenpos.common.Quantity;
 import kitchenpos.menu.domain.*;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.application.OrderService;
-import kitchenpos.common.ControllerTest;
+import kitchenpos.common.ui.ControllerTest;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.dto.OrderLineItemResponse;
